HDDS-5806. remove BackgroundPipelineCreator v1 code
